+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Infrastructure/Repositories/RepositoryBase.cs
@@ -184,8 +184,17 @@ namespace CleanArchitecture.TestApplication.Infrastructure.Repositories
             CancellationToken cancellationToken = default)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            return await dtoProjection.ToListAsync(cancellationToken);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.ToListAsync(cancellationToken);
+        }
+
+        public async Task<TProjection?> FindProjectToAsync<TProjection>(
+            Expression<Func<TPersistence, bool>>? filterExpression,
+            CancellationToken cancellationToken = default)
+        {
+            var queryable = QueryInternal(filterExpression);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.FirstOrDefaultAsync(cancellationToken);
         }
 
         public async Task<IEnumerable> FindAllProjectToWithTransformationAsync<TProjection>(
@@ -194,8 +203,8 @@ namespace CleanArchitecture.TestApplication.Infrastructure.Repositories
             CancellationToken cancellationToken = default)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            var response = transform(dtoProjection);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            var response = transform(projection);
             return await response.Cast<object>().ToListAsync();
         }
 
@@ -205,8 +214,8 @@ namespace CleanArchitecture.TestApplication.Infrastructure.Repositories
             CancellationToken cancellationToken = default)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            var response = filterProjection(dtoProjection);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            var response = filterProjection(projection);
             return await response.Cast<TProjection>().ToListAsync();
         }
     }