+++ Tests/Entities.Constants.TestApplication/Entities.Constants.TestApplication.Infrastructure/Repositories/RepositoryBase.cs
@@ -183,8 +183,17 @@ namespace Entities.Constants.TestApplication.Infrastructure.Repositories
             CancellationToken cancellationToken = default)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            return await dtoProjection.ToListAsync(cancellationToken);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.ToListAsync(cancellationToken);
+        }
+
+        public async Task<TProjection?> FindProjectToAsync<TProjection>(
+            Expression<Func<TPersistence, bool>>? filterExpression,
+            CancellationToken cancellationToken = default)
+        {
+            var queryable = QueryInternal(filterExpression);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.FirstOrDefaultAsync(cancellationToken);
         }
     }
 }