+++ Tests/EntityFramework.SynchronousRepositories/EntityFramework.SynchronousRepositories.Infrastructure/Repositories/RepositoryBase.cs
@@ -251,15 +251,31 @@ namespace EntityFramework.SynchronousRepositories.Infrastructure.Repositories
             CancellationToken cancellationToken = default)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            return await dtoProjection.ToListAsync(cancellationToken);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.ToListAsync(cancellationToken);
+        }
+
+        public async Task<TProjection?> FindProjectToAsync<TProjection>(
+            Expression<Func<TPersistence, bool>>? filterExpression,
+            CancellationToken cancellationToken = default)
+        {
+            var queryable = QueryInternal(filterExpression);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.FirstOrDefaultAsync(cancellationToken);
         }
 
         public List<TProjection> FindAllProjectTo<TProjection>(Expression<Func<TPersistence, bool>>? filterExpression)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            return dtoProjection.ToList();
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return projection.ToList();
+        }
+
+        public TProjection? FindProjectTo<TProjection>(Expression<Func<TPersistence, bool>>? filterExpression)
+        {
+            var queryable = QueryInternal(filterExpression);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return projection.FirstOrDefault();
         }
     }
 }