+++ Tests/GraphQL.AzureFunction.TestApplication/GraphQL.AzureFunction.TestApplication.Infrastructure/Repositories/RepositoryBase.cs
@@ -179,12 +179,21 @@ namespace GraphQL.AzureFunction.TestApplication.Infrastructure.Repositories
         }
 
         public async Task<List<TProjection>> FindAllProjectToAsync<TProjection>(
-            Expression<Func<TPersistence, bool>>? filterExpression,
+            Expression<Func<TPersistence, bool>> filterExpression,
             CancellationToken cancellationToken = default)
         {
             var queryable = QueryInternal(filterExpression);
-            var dtoProjection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
-            return await dtoProjection.ToListAsync(cancellationToken);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.ToListAsync(cancellationToken);
+        }
+
+        public async Task<TProjection> FindProjectToAsync<TProjection>(
+            Expression<Func<TPersistence, bool>> filterExpression,
+            CancellationToken cancellationToken = default)
+        {
+            var queryable = QueryInternal(filterExpression);
+            var projection = queryable.ProjectTo<TProjection>(_mapper.ConfigurationProvider);
+            return await projection.FirstOrDefaultAsync(cancellationToken);
         }
     }
 }