+++ Modules/Intent.Modules.Application.AutoMapper/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 4.0.6
 
 - Improvement: Updated `AutoMapper` NuGet package to `12.0.1` 
+- Improvement: Add an EF Core Interop which adds AutoMapper  Projection Functionality to EF repositoreies.
 
 ### Version 4.0.3